Many changes explained below
Better use of the Factory Enums, no longer need separate enums for the list of units or the list of items.

Projectiles are now all the correct size and are therefore much more visible within the game.

Slightly increase the money added upon defeat of the Boss Enemy to 25.
.
Inline the X and Y co-ordinates of the HUD

Signed-off-by: Jonathon Fitch <lightman1@virginmedia.com>