More changes explained below
Constant classes converted to enums.

Fixed a bug where if Mines depleted the Tower's health but were not colliding with the Tower, the level would not end with a failure.

Fix HUD height for Weapons and move away from getClass().isAssignableFrom in favor of detecting the string "Weapon" in the title.

General refactorings and fixes of warnings + regenerate Javadoc as many comment changes.

Signed-off-by: Jonathon Fitch <lightman1@virginmedia.com>